Set error handler on reader pipe filters and parsers
