+++ src/main/java/org/dita/dost/module/reader/AbstractReaderModule.java
@@ -34,6 +34,7 @@ import javax.xml.namespace.QName;
 import org.apache.commons.io.FileUtils;
 import org.apache.xerces.xni.grammars.XMLGrammarPool;
 import org.dita.dost.exception.DITAOTException;
+import org.dita.dost.exception.DITAOTXMLErrorHandler;
 import org.dita.dost.log.MessageUtils;
 import org.dita.dost.module.AbstractPipelineModuleImpl;
 import org.dita.dost.pipeline.AbstractPipelineInput;
@@ -379,11 +380,14 @@ public abstract class AbstractReaderModule extends AbstractPipelineModuleImpl {
     }
 
     try {
+      final DITAOTXMLErrorHandler errorHandler = new DITAOTXMLErrorHandler(src.toString(), logger, processingMode);
       XMLReader parser = XMLUtils.getXmlReader(ref.format, processingMode).orElse(reader);
+      parser.setErrorHandler(errorHandler);
       XMLReader xmlSource = parser;
       for (final XMLFilter f : getProcessingPipe(currentFile)) {
         f.setParent(xmlSource);
         f.setEntityResolver(CatalogUtils.getCatalogResolver());
+        f.setErrorHandler(errorHandler);
         xmlSource = f;
       }
