+++ src/test/java/org/dita/dost/IntegrationTestPreprocess2.java
@@ -32,16 +32,6 @@ public class IntegrationTestPreprocess2 extends IntegrationTest {
     }
   }
 
-  @Test
-  public void testRngGrammarPoolValidate() throws Throwable {
-    builder()
-      .name("bookmap-rng-based-validate")
-      .transtype(PREPROCESS)
-      .input(Paths.get("main.ditamap"))
-      .put("validate", "true")
-      .test();
-  }
-
   @Disabled
   @Test
   public void testcopyto() throws Throwable {
